Groundstation/requirements.txt
pyserial~=3.4
tensorflow~=2.3.0
numpy~=1.17.2
